suport apollo

+++ sequencer/config.go
@@ -93,5 +93,5 @@ type DBManagerCfg struct {
 	PoolRetrievalInterval    types.Duration `mapstructure:"PoolRetrievalInterval"`
 	L2ReorgRetrievalInterval types.Duration `mapstructure:"L2ReorgRetrievalInterval"`
 	PackBatchWhitelist       []string       `mapstructure:"PackBatchWhitelist"`
-	GasPriceMultiple         uint64         `mapstructure:"GasPriceMultiple"`
+	GasPriceMultiple         float64        `mapstructure:"GasPriceMultiple"`
 }
